requirements.txt
requests
schedule
python-dotenv
supabase
scikit-learn
pandas
fastapi
uvicorn
joblib
pydantic
playwright
